# accounts/openapi.yml
openapi: 3.1.0
info:
  version: 0.1.0
  title: g/accounts
basePath: /accounts
paths:
  /{id}/balance:
    get:
      operationId: Balance
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: integer
        404: {}
  /{id}/transactions:
    get:
      operationId: Transactions
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        200:
          content:
            application/json:
              schema:
                type: array
                oneOf:
                  - $ref: '#/components/schemas/Mint'
                  - $ref: '#/components/schemas/Spend'
                  - $ref: '#/components/schemas/Transfer'
        404: {}
  /:
    post:
      operationId: Register
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required:
                - name
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  accountId:
                    type: integer
                required:
                  - name
        400:
          content:
            text/plain:
              schema:
                type: string
  /{id}/mint:
    post:
      operationId: Mint
      parameters:
        - $ref: '#/components/parameters/AccountId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: integer
              required:
                - amount
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactionId:
                    type: integer
        400:
          content:
            text/plain:
              schema:
                type: string
        404: {}
  /{id}/spend:
    post:
      operationId: Spend
      parameters:
        - $ref: '#/components/parameters/AccountId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: integer
              required:
                - amount
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactionId:
                    type: integer
        400:
          content:
            text/plain:
              schema:
                type: string
        404: {}
  /{id}/transfer:
    post:
      operationId: Transfer
      parameters:
        - $ref: '#/components/parameters/AccountId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: integer
                to:
                  type: integer
              required:
                - amount
                - to
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactionId:
                    type: integer
        400:
          content:
            text/plain:
              schema:
                type: string
        404: {}
components:
  parameters:
    AccountId:
      in: path
      name: id
      schema:
        type: integer
      required: true
  schemas:
    Transaction:
      type: object
      properties:
        id:
          type: string
    Mint:
      type: object
      properties:
        type:
          type: string
          enum: ["mint"]
        inserted_at:
          type: string
          format: date-time
        amount:
          type: integer
    Spend:
      type: object
      properties:
        type:
          type: string
          enum: ["spend"]
        inserted_at:
          type: string
          format: date-time
        amount:
          type: integer
    Transfer:
      type: object
      properties:
        type:
          type: string
          enum: ["transfer"]
        inserted_at:
          type: string
          format: date-time
        amount:
          type: integer
        recipient:
          type: integer